+++ server/engine/bomb.go
@@ -9,7 +9,7 @@ import (
 type Bomb struct {
 	id     uint8
 	evil   bool
-	damage uint16
+	damage float32
 	x      float32
 	y      float32
 	origin *Point
@@ -37,7 +37,7 @@ func (bomb *Bomb) Tick(delta time.Duration) {
 	bomb.timer -= float32(delta) / float32(time.Second)
 }
 
-func DefaultBomb(id uint8, x float32, y float32, origin Object, evil bool, damage uint16, timer float32) *Bomb {
+func DefaultBomb(id uint8, x float32, y float32, origin Object, evil bool, damage float32, timer float32) *Bomb {
 	return &Bomb{
 		id:     id,
 		evil:   evil,
